refactor : 테스트 코드를 MVC패턴화 함에 따른 패키지 분리. MatchResultTest 생성 및 LottoTest에서 MatchResult관련 메서드 옮김
