+++ src/main/java/lotto/Buyer.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto;
 
 import lotto.controller.LottoGenerator;
