+++ src/main/java/lotto/LottoTicket.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto;
 
 import java.util.ArrayList;
 import java.util.HashSet;
