+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -1,7 +1,7 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.LottoNumber;
-import lotto.domain.LottoTicket;
+import lotto.LottoNumber;
+import lotto.LottoTicket;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
